feat: use server-side apply in the KAS bootstrap container
Replace CreateOrUpdate with server-side apply for applying bootstrap
manifests. CreateOrUpdate cannot apply resources generically as all
resource types need to be special cased with a callback function.

Introduce a two-client architecture: an admin (localhost) client applies
the kas-bootstrap-container ClusterRoleBinding first to grant
cluster-admin, then a second client using the kas-bootstrap-container
identity applies the remaining manifests.

The FeatureGate CR has a ValidatingAdmissionPolicy that restricts spec
changes to the system:hosted-cluster-config user, so
reconcileFeatureGate
now uses the kas-bootstrap-container client and this user is also added
to the ValidatingAdmissionPolicy so it can apply the FeatureGate spec.